+++ cmd/agt/main.go
@@ -92,6 +92,8 @@ func run(args []string, stdout, stderr io.Writer) int {
 		return cmdNetguard(args[1:], stdout, stderr)
 	case "ratelimit":
 		return cmdRateLimit(args[1:], stdout, stderr)
+	case "webhook":
+		return cmdWebhook(args[1:], stdout, stderr)
 	case "doctor":
 		return cmdDoctor(args[1:], stdout, stderr)
 	case "quickstart":